requirements.txt
scrapy>=2.11,<3
scrapy-playwright>=0.0.34
playwright>=1.40
PyYAML>=6.0.1
psycopg[binary]>=3.2.3
python-dotenv>=1.0.1
requests>=2.31.0
